refact: move getToken() to lib/transferwise
